Beta 3 fixes, cleanup, and accessibility previews
